Migrate validationUtils to formUtils
